requirements.txt
appdirs==1.4.3
Django==1.9
django-filter==1.0.2
djangorestframework==3.6.3
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
Markdown==2.6.8
odfpy==1.3.4
packaging==16.8
pyparsing==2.2.0
six==1.10.0
texttable==0.8.8
whitenoise==3.3.0
xlrd==1.0.0
xlwt==1.2.0
